api: move route logic to controllers
